feat(advertisements): complete advertisement CRUD API with tests

+++ app/models/advertisement.py
@@ -1,5 +1,5 @@
 from app.db.database import Base
-from sqlalchemy import Column, Integer, String, Numeric, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Float
 from datetime import datetime
 from sqlalchemy.orm import relationship
 
@@ -8,7 +8,7 @@ class Advertisement(Base):
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String)
     description = Column(String)
-    price = Column(Numeric)
+    price = Column(Float)
     user_id = Column(Integer, ForeignKey("users.id"))
     created_at = Column(DateTime, default=datetime.now)
     updated_at = Column(DateTime, default=datetime.now)
